# internal/transport/http/swaggerdocs/swagger.yaml
basePath: /api/v1
definitions:
...
  shared.BondsHonorCharacterUnitResponse:
    properties:
      gameCharacterId:
        type: integer
      id:
        type: integer
      unit:
        type: string
    type: object
  shared.BondsHonorGroupResponse:
    properties:
      characterId1:
        type: integer
      characterId2:
        type: integer
      groupId:
        type: integer
    type: object
  shared.BondsHonorLevelResponse:
    properties:
      bondsHonorId:
        type: integer
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
    type: object
  shared.BondsHonorListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/shared.BondsHonorObjectResponse'
        type: array
      pagination:
        $ref: '#/definitions/shared.PaginationResponse'
    type: object
  shared.BondsHonorObjectResponse:
    properties:
      bondsGroup:
        $ref: '#/definitions/shared.BondsHonorGroupResponse'
      bondsGroupId:
        type: integer
      characterUnit1:
        $ref: '#/definitions/shared.BondsHonorCharacterUnitResponse'
      characterUnit2:
        $ref: '#/definitions/shared.BondsHonorCharacterUnitResponse'
      configurableUnitVirtualSinger:
        type: boolean
      description:
        type: string
      gameCharacterUnitId1:
        type: integer
      gameCharacterUnitId2:
        type: integer
      honorRarity:
        type: string
      id:
        type: integer
      levels:
        items:
          $ref: '#/definitions/shared.BondsHonorLevelResponse'
        type: array
      name:
        type: string
      pronunciation:
        type: string
      seq:
        type: integer
    type: object
...
  shared.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  shared.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/shared.ErrorBody'
    type: object
...
  shared.PaginationResponse:
    properties:
      has_next:
        type: boolean
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for master data sync and card querying.
  title: sekai-master-api
  version: "1.0"
paths:
...
  /bondsHonors/{region}/{id}:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: Bonds honor ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.BondsHonorObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Get a bonds honor by id
      tags:
      - bondsHonors
  /bondsHonors/{region}/list:
    get:
      parameters:
      - description: Region
        in: path
        name: region
        required: true
        type: string
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: Sort field
        enum:
        - id
        - seq
        - bondsGroupId
        - gameCharacterUnitId1
        - gameCharacterUnitId2
        - honorRarity
        - name
        in: query
        name: sort_by
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Exact bonds group ID
        in: query
        minimum: 1
        name: bonds_group_id
        type: integer
      - description: Exact first game character unit ID
        in: query
        minimum: 1
        name: game_character_unit_id1
        type: integer
      - description: Exact second game character unit ID
        in: query
        minimum: 1
        name: game_character_unit_id2
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.BondsHonorListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: List bonds honors by page
      tags:
      - bondsHonors
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"